requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
pandas>=1.3.3
matplotlib>=3.4.3
opencv-python>=4.5.3
scikit-learn>=1.0.0
tqdm>=4.62.3
Pillow>=8.3.2
albumentations>=1.1.0
tensorboard>=2.7.0
PyYAML>=5.4.1
psutil>=5.8.0 